Add files via upload
The R code of this project focuses on the care economy and data analysis of the National Time Use Survey (ENUT) in Colombia. The objective of the project is to analyze and study how time is distributed among different caregiving activities in the Colombian population.

The code performs various tasks to achieve this objective. Firstly, it loads the necessary libraries for data analysis in R. Then, it reads the data files in DTA format, which contain the information collected in the ENUT.

Next, the code merges the data sets to have a more comprehensive view of caregiving activities. It performs data transformations to ensure data quality and coherence. Additionally, it calculates variables related to indirect and direct care, enabling the analysis of how time is distributed between these activities.

Finally, the code conducts descriptive analysis of the data, providing descriptive statistics that help to better understand the distribution and patterns of time dedicated to care in the Colombian population.